+++ cli_options/CHANGELOG.md
@@ -4,6 +4,34 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+* Support providing repeating arguments with a separator. If you set the `separator`
+  option for an argument's schema you can pass the values in the format `--arg value1<sep>value2`.
+  For example, for the following schema:
+
+  ```elixir
+  schema = [
+    name: [
+      type: :string,
+      multiple: true,
+      separator: ";" 
+    ]
+  ```
+
+  all of the following invocation are valid and equivalent:
+
+  ```
+  # passing the arg multiple times  
+  $ mix foo --name john --name jack --name paul
+
+  # passing the arg once with the values separated with ;
+  $ mix foo --name john;jack;paul
+
+  # a combination of the above
+  $ mix foo --name john --name jack;paul
+  ```
+
 ## [v0.1.2](https://github.com/sportradar/elixir-workspace/tree/cli_options/v0.1.2) (2024-07-12)
 
 ### Added